add search and back button
